Wired modal into component

+++ src/app/courses-list/courses-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Course } from './course.model';
 import { CoursesService } from './courses.service';
+import { ModalService } from '../services/modal.service';
 
 @Component({
   selector: 'app-courses-list',
@@ -11,14 +12,16 @@ export class CoursesListComponent implements OnInit {
   courses: Course[] = [];
   orderByDate = (a: Course, b: Course) => (a.creationDate < b.creationDate ? -1 : (a.creationDate > b.creationDate ? 1 : 0));
 
-  constructor(private coursesService: CoursesService) { }
+  constructor(private coursesService: CoursesService, private modalService: ModalService) { }
 
   ngOnInit() {
     this.courses = this.coursesService.getCourses();
   }
 
   onDelete(id: number) {
-    this.coursesService.removeCourse(id);
+    this.modalService.showModal('Are you sure you want to delete?').subscribe(() => {
+      this.coursesService.removeCourse(id);
+    });
   }
 
 }
